trading_engine/requirements.txt
ccxt
python-dotenv
fastapi
redis
types-redis
uvicorn
Jinja2Templates